First draft of the PyCRAMGym class
